requirements-dev-contrib.txt
black==22.1.0
flake8==3.8.3
isort==5.4.2
pre-commit>=2.6.0
pytest-cov>=2.8.1
pytest-order>=0.9.5
pyupgrade==2.7.2
